# chore: Improve benchmarks