--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -737,3 +737,40 @@ jobs:
         run: make prepublish
       - name: Publint
         run: node ./scripts/publint.ts
+
+  codspeed:
+    name: Codspeed
+    runs-on: ubuntu-latest
+    needs: prepare-yarn-cache
+    steps:
+      - name: Free Disk Space (Ubuntu)
+        uses: jlumbroso/free-disk-space@main
+        with:
+          # this might remove tools that are actually needed,
+          # if set to "true" but frees about 6 GB
+          tool-cache: false
+          # all of these default to true, but feel free to set to
+          # "false" if necessary for your workflow
+          android: true
+          dotnet: true
+          haskell: true
+          large-packages: true
+          docker-images: true
+          swap-storage: true
+      - name: Checkout code
+        uses: actions/checkout@v4
+      - name: Use Node.js latest
+        uses: actions/setup-node@v4
+        with:
+          node-version: latest
+          check-latest: true
+          cache: "yarn"
+      - name: Install
+        run: yarn install
+      - name: Build
+        run: make prepublish-build
+      - name: Run benchmarks
+        uses: CodSpeedHQ/action@v4
+        with:
+          mode: walltime
+          run: node --expose-gc ./benchmark/index.mjs --only-current