--- a/Gulpfile.ts
+++ b/Gulpfile.ts
@@ -1197,3 +1197,26 @@ gulp.task(
   "watch",
   process.env.WATCH_SKIP_BUILD ? watch : gulp.series("build-dev", watch)
 );
+
+gulp.task("build-perf", function () {
+  const pkgs = ["packages/babel-parser"].map(
+    src =>
+      ({
+        src,
+        format: "esm",
+        dest: "lib",
+        inputs: [getIndexFromPackage(src)],
+      }) as PackageInfo
+  );
+  return buildRollup(pkgs);
+});
+
+gulp.task(
+  "watch-perf",
+  gulp.series("build-perf", function watchPerf() {
+    if (!process.env.IS_PUBLISH) {
+      throw new Error("This task must be run with `process.env.IS_PUBLISH`");
+    }
+    gulp.watch(defaultSourcesGlob, gulp.task("build-perf"));
+  })
+);