--- a/Makefile
+++ b/Makefile
@@ -61,6 +61,9 @@ build-plugin-transform-runtime-dist:
 watch:
 	$(MAKEJS) watch
 
+watch-perf:
+	$(MAKEJS) watch-perf
+
 code-quality: lint
 
 tscheck: