--- a/benchmark/all/real-case-js-node6.mjs
+++ b/benchmark/all/real-case-js-node6.mjs
@@ -1,0 +1,40 @@
+import {
+  loadFixtures,
+  generateCaseName,
+  baselineCore,
+  baselinePresetEnv,
+  currentPresetEnv,
+  currentCore,
+  addBenchCase,
+} from "../util.mjs";
+
+const opts = {
+  filename: "file.ts",
+  configFile: false,
+  babelrc: false,
+  browserslistConfigFile: false,
+  generatorOpts: {
+    compact: false,
+  },
+  targets: "node 6",
+};
+
+const fixtures = loadFixtures("js");
+
+fixtures.forEach(({ name, content }) => {
+  addBenchCase(
+    `${generateCaseName(import.meta.url)} ${name}`,
+    () => {
+      baselineCore.transformSync(content, {
+        presets: [baselinePresetEnv],
+        ...opts,
+      });
+    },
+    () => {
+      currentCore.transformSync(content, {
+        presets: [currentPresetEnv],
+        ...opts,
+      });
+    }
+  );
+});