--- a/benchmark/babel-parser/real-case.mjs
+++ b/benchmark/babel-parser/real-case.mjs
@@ -1,0 +1,24 @@
+import {
+  loadFixtures,
+  addBenchCase,
+  baselineParser,
+  currentParser,
+  generateCaseName,
+} from "../util.mjs";
+
+const fixtures = loadFixtures();
+
+fixtures.forEach(({ name, content }) => {
+  addBenchCase(
+    `${generateCaseName(import.meta.url)} ${name}`,
+    baselineParser.parse,
+    currentParser.parse,
+    [
+      content,
+      {
+        sourceType: "module",
+        plugins: name.endsWith(".ts") ? ["typescript"] : [],
+      },
+    ]
+  );
+});