--- a/benchmark/index.mjs
+++ b/benchmark/index.mjs
@@ -1,0 +1,6 @@
+import "./babel-parser/real-case.mjs";
+import "./babel-traverse/real-case.mjs";
+import "./babel-generator/real-case.mjs";
+import "./all/real-case-ts-mjs.mjs";
+import "./all/real-case-mjs-cjs.mjs";
+import "./all/real-case-js-node6.mjs";