--- a/benchmark/package.json
+++ b/benchmark/package.json
@@ -4,14 +4,17 @@
   "type": "module",
   "devDependencies": {
     "$repo-utils": "link:../scripts/repo-utils",
-    "@babel-baseline/cli": "npm:@babel/cli@7.27.1",
-    "@babel-baseline/core": "npm:@babel/core@7.24.4",
-    "@babel-baseline/generator": "npm:@babel/generator@7.24.4",
-    "@babel-baseline/helper-compilation-targets": "npm:@babel/helper-compilation-targets@7.23.6",
-    "@babel-baseline/helper-validator-identifier": "npm:@babel/helper-validator-identifier@7.22.20",
-    "@babel-baseline/parser": "npm:@babel/parser@7.24.4",
-    "@babel-baseline/traverse": "npm:@babel/traverse@7.24.1",
-    "@babel-baseline/types": "npm:@babel/types@7.24.0",
+    "@babel-baseline/cli": "npm:@babel/cli@8.0.0-rc.3",
+    "@babel-baseline/core": "npm:@babel/core@8.0.0-rc.3",
+    "@babel-baseline/generator": "npm:@babel/generator@8.0.0-rc.3",
+    "@babel-baseline/helper-compilation-targets": "npm:@babel/helper-compilation-targets@8.0.0-rc.3",
+    "@babel-baseline/helper-validator-identifier": "npm:@babel/helper-validator-identifier@8.0.0-rc.3",
+    "@babel-baseline/parser": "npm:@babel/parser@8.0.0-rc.3",
+    "@babel-baseline/plugin-transform-modules-commonjs": "npm:@babel/plugin-transform-modules-commonjs@8.0.0-rc.3",
+    "@babel-baseline/preset-env": "npm:@babel/preset-env@8.0.0-rc.3",
+    "@babel-baseline/preset-typescript": "npm:@babel/preset-typescript@8.0.0-rc.3",
+    "@babel-baseline/traverse": "npm:@babel/traverse@8.0.0-rc.3",
+    "@babel-baseline/types": "npm:@babel/types@8.0.0-rc.3",
     "@babel/cli": "workspace:^",
     "@babel/core": "workspace:^",
     "@babel/generator": "workspace:^",
@@ -24,7 +27,10 @@
     "@babel/preset-typescript": "workspace:^",
     "@babel/traverse": "workspace:^",
     "@babel/types": "workspace:^",
-    "benchmark": "^2.1.4"
+    "@codspeed/tinybench-plugin": "patch:@codspeed/tinybench-plugin@npm%3A5.3.0#~/.yarn/patches/@codspeed-tinybench-plugin-npm-5.3.0-479a7eef48.patch",
+    "benchmark": "^2.1.4",
+    "glob": "^10.3.10",
+    "tinybench": "^6.0.0"
   },
   "version": "8.0.0-rc.6"
 }