--- a/benchmark/util.mjs
+++ b/benchmark/util.mjs
@@ -1,17 +1,56 @@
 import * as currentTypes from "@babel/types";
 import * as baselineTypes from "@babel-baseline/types";
-import * as currentParser from "@babel/parser";
-import * as baselineParser from "@babel-baseline/parser";
-import * as currentGenerator from "@babel/generator";
-import * as baselineGenerator from "@babel-baseline/generator";
+import currentGenerator from "@babel/generator";
+import baselineGenerator from "@babel-baseline/generator";
 import * as currentCore from "@babel/core";
 import * as baselineCore from "@babel-baseline/core";
-import * as currentTraverse from "@babel/traverse";
-import * as baselineTraverse from "@babel-baseline/traverse";
-import presetEnv from "@babel/preset-env";
-import presetTypescript from "@babel/preset-typescript";
+import currentTraverse from "@babel/traverse";
+import baselineTraverse from "@babel-baseline/traverse";
+import currentPresetEnv from "@babel/preset-env";
+import currentPresetTypescript from "@babel/preset-typescript";
+import currentPluginCommonjs from "@babel/plugin-transform-modules-commonjs";
+import baselinePresetEnv from "@babel-baseline/preset-env";
+import baselinePresetTypescript from "@babel-baseline/preset-typescript";
+import baselinePluginCommonjs from "@babel-baseline/plugin-transform-modules-commonjs";
 
-import benchmark from "benchmark";
+import { Bench } from "tinybench";
+import { globSync } from "glob";
+import { withCodSpeed } from "@codspeed/tinybench-plugin";
+import { readFileSync, copyFileSync, mkdirSync } from "node:fs";
+import { fileURLToPath, pathToFileURL } from "node:url";
+import path from "node:path";
+import assert from "node:assert";
+import { tmpdir } from "node:os";
+
+const currentParser = await copyPkg("@babel/parser");
+const baselineParser = await copyPkg("@babel-baseline/parser");
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const onlyCurrent = process.argv.includes("--only-current");
+const onlyBaseline = process.argv.includes("--only-baseline");
+
+if (!globalThis.gc) {
+  console.warn("Recommend running with --expose-gc.");
+}
+
+async function copyPkg(name) {
+  const src = fileURLToPath(import.meta.resolve(name));
+  const dst = path.join(
+    tmpdir(),
+    "babel-benchmark",
+    name.replace("/", "-") + ".js"
+  );
+  mkdirSync(path.dirname(dst), { recursive: true });
+  copyFileSync(src, dst);
+  try {
+    return await import(pathToFileURL(dst));
+  } catch (error) {
+    console.error("Please run with `make prepublish-build`");
+    return await import(name);
+  }
+}
 
 export function report(event) {
   const bench = event.target;
@@ -31,21 +70,105 @@ function formatNumber(x) {
   return `${Math.round(x)}`.replace(/\d(?=(?:\d{3})+$)/g, "$&_");
 }
 
+function loadFixtures(type = "*") {
+  const fixtures = globSync(`fixtures/*.${type}.txt`, {
+    cwd: __dirname,
+  }).sort();
+
+  return fixtures.map(fixture => ({
+    name: path.basename(fixture, ".txt"),
+    content: readFileSync(path.join(__dirname, fixture), "utf-8"),
+  }));
+}
+
+function generateCaseName(url) {
+  const filename = fileURLToPath(url);
+  return path.relative(__dirname, filename).replace(/\\/g, "/");
+}
+
 class Benchmark {
   constructor() {
-    this.suite = new benchmark.Suite();
+    this.bench = withCodSpeed(
+      new Bench({
+        time: 5000,
+        warmup: false,
+        retainSamples: true,
+      })
+    );
   }
   add(name, fn, options) {
-    this.suite.add(name, fn, options);
+    this.bench.add(name, fn, options);
   }
-  run() {
-    this.suite.on("cycle", report);
-    this.suite.on("error", function (event) {
-      console.error(event.target.error);
+  async run() {
+    this.bench.addEventListener("cycle", function (event) {
+      const task = event.task;
+      const { latency, throughput } = task.result;
+      const timeMs = latency.mean;
+      const time =
+        timeMs < 10
+          ? `${Math.round(timeMs * 1000) / 1000}ms`
+          : `${Math.round(timeMs)}ms`;
+      const msg = `${task.name}: ${formatNumber(throughput.mean)} ops/sec ±${
+        Math.round(latency.rme * 100) / 100
+      }% ${latency.samplesCount} runs (${time})`;
+      console.log(msg);
+
+      globalThis.gc?.();
     });
-    this.suite.run();
+    this.bench.addEventListener("error", function (event) {
+      console.error(event.task.result.error);
+    });
+    this.bench.runSync();
   }
 }
+
+const bench = new Benchmark();
+
+const baselineCases = [];
+const currentCases = [];
+
+function addBaselineCase(name, fn, args) {
+  baselineCases.push({
+    name: `${name} @ baseline`,
+    fn: () => {
+      fn.apply(null, args);
+    },
+  });
+}
+
+function addCurrentCase(name, fn, args) {
+  currentCases.push({
+    name: `${name} @ current`,
+    fn: () => {
+      fn.apply(null, args);
+    },
+  });
+}
+
+function addBenchCase(name, baseline, current, args) {
+  addBaselineCase(name, baseline, args);
+  addCurrentCase(name, current, args);
+}
+
+setTimeout(async () => {
+  if (onlyCurrent) {
+    currentCases.forEach(({ name, fn }) => {
+      bench.add(name, fn, { async: false });
+    });
+  } else if (onlyBaseline) {
+    baselineCases.forEach(({ name, fn }) => {
+      bench.add(name, fn, { async: false });
+    });
+  } else {
+    assert.strictEqual(baselineCases.length, currentCases.length);
+    for (let i = 0; i < baselineCases.length; i++) {
+      bench.add(currentCases[i].name, currentCases[i].fn, { async: false });
+      bench.add(baselineCases[i].name, baselineCases[i].fn, { async: false });
+    }
+  }
+
+  bench.run();
+}, 0);
 
 export {
   currentTypes,
@@ -58,7 +181,17 @@ export {
   baselineCore,
   currentTraverse,
   baselineTraverse,
-  presetEnv,
-  presetTypescript,
+  currentPresetEnv,
+  currentPresetTypescript,
+  currentPluginCommonjs,
+  baselinePresetEnv,
+  baselinePresetTypescript,
+  baselinePluginCommonjs,
   Benchmark,
+  loadFixtures,
+  generateCaseName,
+  bench as benchmark,
+  addBenchCase,
+  addBaselineCase,
+  addCurrentCase,
 };