… Makefile.js
… benchmark/babel-generator/real-case/jquery-3.6.txt
… benchmark/babel-generator/real-case/jquery.mjs
… benchmark/babel-parser/real-case-ts/bench.mjs
… benchmark/babel-parser/real-case-ts/ts-parser.txt
… benchmark/babel-parser/real-case/bench.mjs
… benchmark/babel-parser/real-case/jquery-3.6.txt
… benchmark/fixtures/babel-parser-express.ts.txt
… benchmark/fixtures/jquery-3.6.js.txt
… benchmark/fixtures/ts-checker.mjs.txt
… benchmark/fixtures/ts-checker.ts.txt
… benchmark/fixtures/ts-parser.mjs.txt
… benchmark/fixtures/ts-parser.ts.txt
… yarn.lock