@nicolo-ribaudo 2024-09-30
> Perhaps we can also use [codspeed.io](https://codspeed.io/), I recently noticed that some repositories started using it, and it looks good. :)

If you want to try setting it up, it seems ok 👍