@JLHwung 2024-10-01
> there is some unknown reason why the baseline parser is faster.

I can recall this happened before. Will the baseline still be faster if we run baseline vs main on different node process?