@liuxingbaoyu 2024-10-01
> I can recall this happened before. Will the baseline still be faster if we run baseline vs main on different node process?

I split the run into two and still the baseline was faster. :)