@liuxingbaoyu 2024-10-01
> What if you build current using `make prepublish-build`, so that it removes all the `BABEL_8_BREAKING` code?

Yes, I'm using `make prepublish-build`, otherwise it's a lot slower. :)