@liuxingbaoyu 2024-10-14
https://github.com/babel/babel/blob/33d705e8df0ec660c1e8e67154949d2d89269350/packages/babel-parser/src/plugins/estree.ts#L19-L20
Commenting them out seems to give a 2x performance boost.🤔
![image](https://github.com/user-attachments/assets/022931d8-332b-4074-acdb-04973b27a08a)