@nicolo-ribaudo 2024-10-14
ESTree doesn't have that `.index` property and that's why we are hiding it. However, we could say that in Babel 8 that's another place where we diverge from the ESTree spec and leave it there.