@liuxingbaoyu 2024-10-14
![image](https://github.com/user-attachments/assets/ee290ddb-aa47-4d15-8548-5b35cbad336b)
I'm surprised this would slow down by 3x, which means `useDefineForClassFields` may inadvertently bring performance impacts.