@nicolo-ribaudo 2024-10-14
How is the perf with a Babel 8 build, where we do not transpile class fields?