@liuxingbaoyu 2024-10-14
We don't actually use this in Babel 7 currently, this was my attempt to emulate the new TS behavior in the hope of getting better optimizations from V8, unfortunately to no avail.