@liuxingbaoyu 2024-10-14
![image](https://github.com/user-attachments/assets/0bad369f-18c6-49ae-b943-3c2d96ee0b96)
Also interestingly, bun is faster than node here. :)