@liuxingbaoyu 2024-10-18
I have a local branch that tries to change all `Position` to `number` to introduce a mode without `node.loc` to improve performance, which can also solve the estree problem.
But unfortunately the performance of the normal mode has dropped by 10%.
I am trying a new approach.