@liuxingbaoyu 2024-10-18
Haha, I seem to have found something interesting!
When we add any property here, the performance of the parser will drop by about 10%.
https://github.com/babel/babel/blob/d65873827b00e4a0a3ed8fe59000cebd5d1dd82e/packages/babel-parser/src/options.ts#L8-L28
This also happens in https://github.com/babel/babel/pull/16849.
before
```
PS F:\babel\benchmark\babel-parser\real-case> node bench.mjs
current 4 jquery 3.6: 24.85 ops/sec ±2.32% (40ms)
baseline 4 jquery 3.6: 24.65 ops/sec ±1.7% (41ms)
current 4 jquery 3.6: 24.47 ops/sec ±1.33% (41ms)
baseline 4 jquery 3.6: 25.39 ops/sec ±1.34% (39ms)
```
after
```
PS F:\babel\benchmark\babel-parser\real-case> node bench.mjs
current 4 jquery 3.6: 22.69 ops/sec ±2.52% (44ms)
baseline 4 jquery 3.6: 25.61 ops/sec ±1.96% (39ms)
current 4 jquery 3.6: 22.91 ops/sec ±1.51% (44ms)
baseline 4 jquery 3.6: 25.2 ops/sec ±1.69% (40ms)
```

I currently suspect that the number of properties reaches 20, causing v8 to convert it into a slow object.
I will try to open a PR.