@liuxingbaoyu 2025-12-19
The parser may not have performance regression.
This could be due to performance instability in GitHub CI.